backend/requirements.txt
fastapi
uvicorn
pandas
openpyxl
python-dotenv
pydantic
python-multipart
anthropic[vertex]
tavily-python
firecrawl-py
